requirements.txt
art==4.6
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.15
django-allauth==0.42.0
django-appconf==1.0.4
django-bootstrap-pagination==1.7.1
django-bootstrap4==1.1.1
django-cors-headers==3.3.0
django-extensions==2.2.9
django-heroku==0.3.1
django-imagekit==4.0.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
drf-yasg==1.17.1
Faker==4.1.0
gunicorn==20.0.4
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
pilkit==2.0
Pillow==7.1.2
psycopg2==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
soupsieve==2.0.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
